+++ TidalPy/RadialSolver/starting/takeuchi.pyx
@@ -36,78 +36,70 @@ cdef void cf_takeuchi_solid_dynamic_compressible(
 
     Parameters
     ----------
-    radius : float
-        Radius where the radial functions are calculated. [m]
-    shear_modulus : Union[float, complex]
-        Shear modulus (can be complex for dissipation) at `radius` [Pa]
-    bulk_modulus : Union[float, complex]
-        Bulk modulus (can be complex for dissipation) at `radius` [Pa]
-    density : float
-        Density at `radius` [kg m-3]
-    frequency : float
+    frequency : double
         Forcing frequency (for spin-synchronous tides this is the orbital motion) [rad s-1]
-    degree_l : int = 2
+    radius : double
+        Radius where the radial functions are calculated. [m]
+    density : double
+        Density at `radius` [kg m-3]
+    bulk_modulus : double complex
+        Bulk modulus (can be complex for dissipation) at `radius` [Pa]
+    shear_modulus : double complex
+        Shear modulus (can be complex for dissipation) at `radius` [Pa]
+    degree_l : unsigned char
         Tidal harmonic order.
-    G_to_use : float = G
+    G_to_use : double
         Gravitational constant. Provide a non-dimensional version if the rest of the inputs are non-dimensional.
-
-    Returns
-    -------
-    solid_guesses : np.ndarray
-        The three independent solid guesses (sn1, sn2, sn3)
+    num_ys : ssize_t
+        Number of radial solutions for this layer type.
+    starting_conditions_ptr : double complex*, <Output>
+        Desired starting conditions for this layer.
+        Three independent solid guesses (sn1, sn2, sn3)
 
     """
 
     # Convert compressibility parameters
-    cdef double complex lame
-    lame = bulk_modulus - (2. / 3.) * shear_modulus
+    cdef double complex lame = bulk_modulus - (2. / 3.) * shear_modulus
 
-    # Constants (See Eqs. B13-B16 of KMN15)
-    cdef double dynamic_term, gamma
-    cdef double complex alpha2, beta2
-    dynamic_term = frequency * frequency
-    alpha2       = (lame + 2. * shear_modulus) / density
-    beta2        = shear_modulus / density
-    gamma        = 4. * pi * G_to_use * density / 3.
+    # Constants
+    cdef double gamma           = 4. * pi * G_to_use * density / 3.
+    cdef double dynamic_term    = frequency * frequency
+    cdef double complex alpha2  = (lame + 2. * shear_modulus) / density
+    cdef double complex beta2   = shear_modulus / density
 
     # Optimizations
-    cdef double r_inverse, r2, degree_l_dbl, lp1, lm1, dlp1, llp1, dlp3
-    r_inverse    = 1. / radius
-    r2           = radius * radius
-    degree_l_dbl = <double> degree_l
-    lp1          = degree_l_dbl + 1.
-    lm1          = degree_l_dbl - 1.
-    dlp1         = 2. * degree_l_dbl + 1.
-    dlp3         = 2. * degree_l_dbl + 3.
-    llp1         = degree_l_dbl * lp1
+    cdef double r_inverse    = 1. / radius
+    cdef double r2           = radius * radius
+    cdef double degree_l_dbl = <double> degree_l
+    cdef double lp1          = degree_l_dbl + 1.
+    cdef double lm1          = degree_l_dbl - 1.
+    cdef double dlp1         = 2. * degree_l_dbl + 1.
+    cdef double dlp3         = 2. * degree_l_dbl + 3.
+    cdef double llp1         = degree_l_dbl * lp1
 
     # Helper functions
-    cdef double complex k2_quad_pos, k2_quad_neg, k2_quad, k2_quad_sqrt
-    k2_quad_pos  = (dynamic_term / beta2) + ((dynamic_term + 4. * gamma) / alpha2)
-    k2_quad_neg  = (dynamic_term / beta2) - ((dynamic_term + 4. * gamma) / alpha2)
-    k2_quad      = (k2_quad_neg * k2_quad_neg) + ((4. * llp1 * gamma**2) / (alpha2 * beta2))
-    k2_quad_sqrt = cf_csqrt(k2_quad)
+    # See Eq. 99 of TS72
+    cdef double complex k2_quad_pos  = (dynamic_term / beta2) + ((dynamic_term + 4. * gamma) / alpha2)
+    cdef double complex k2_quad_neg  = (dynamic_term / beta2) - ((dynamic_term + 4. * gamma) / alpha2)
+    cdef double complex k2_quad      = (k2_quad_neg * k2_quad_neg) + ((4. * llp1 * gamma**2) / (alpha2 * beta2))
+    cdef double complex k2_quad_sqrt = cf_csqrt(k2_quad)
 
     # QUESTION: (Issue #43) TS74 has these flipped compared to KMN15. Going with TS74 for this func.
     # See the -/+ order in TS72 EQ. 99
     cdef size_t neg_index = 0
     cdef size_t pos_index = 1
-    cdef double complex k2_pos, k2_neg
-    k2_neg = (1. / 2.) * (k2_quad_pos - k2_quad_sqrt)
-    k2_pos = (1. / 2.) * (k2_quad_pos + k2_quad_sqrt)
+    cdef double complex k2_neg = (1. / 2.) * (k2_quad_pos - k2_quad_sqrt)
+    cdef double complex k2_pos = (1. / 2.) * (k2_quad_pos + k2_quad_sqrt)
 
-    cdef double complex f_pos, f_neg
-    f_pos = (beta2 * k2_pos - dynamic_term) / gamma
-    f_neg = (beta2 * k2_neg - dynamic_term) / gamma
+    cdef double complex f_pos = (beta2 * k2_pos - dynamic_term) / gamma
+    cdef double complex f_neg = (beta2 * k2_neg - dynamic_term) / gamma
 
-    cdef double complex h_pos, h_neg
-    h_pos = f_pos - lp1
-    h_neg = f_neg - lp1
+    cdef double complex h_pos = f_pos - lp1
+    cdef double complex h_neg = f_neg - lp1
 
     # Calculate Takeuchi and Saito functions
-    cdef double complex z2_pos, z2_neg
-    z2_pos = k2_pos * r2
-    z2_neg = k2_neg * r2
+    cdef double complex z2_pos = k2_pos * r2
+    cdef double complex z2_neg = k2_neg * r2
 
     cdef double complex phi_pos, phi_lp1_pos, psi_pos
     cdef double complex phi_neg, phi_lp1_neg, psi_neg
@@ -206,74 +198,65 @@ cdef void cf_takeuchi_solid_static_compressible(
 
     Parameters
     ----------
-    radius : float
+    radius : double
         Radius where the radial functions are calculated. [m]
-    shear_modulus : Union[float, complex]
-        Shear modulus (can be complex for dissipation) at `radius` [Pa]
-    bulk_modulus : Union[float, complex]
-        Bulk modulus (can be complex for dissipation) at `radius` [Pa]
-    density : float
+    density : double
         Density at `radius` [kg m-3]
-    degree_l : int = 2
+    bulk_modulus : double complex
+        Bulk modulus (can be complex for dissipation) at `radius` [Pa]
+    shear_modulus : double complex
+        Shear modulus (can be complex for dissipation) at `radius` [Pa]
+    degree_l : unsigned char
         Tidal harmonic order.
-    G_to_use : float = G
+    G_to_use : double
         Gravitational constant. Provide a non-dimensional version if the rest of the inputs are non-dimensional.
-
-    Returns
-    -------
-    solid_guesses : np.ndarray
-        The three independent solid guesses (sn1, sn2, sn3)
-
+    num_ys : ssize_t
+        Number of radial solutions for this layer type.
+    starting_conditions_ptr : double complex*, <Output>
+        Desired starting conditions for this layer.
+        Three independent solid guesses (sn1, sn2, sn3)
     """
 
     # Convert compressibility parameters
-    cdef double complex lame
-    lame = bulk_modulus - (2. / 3.) * shear_modulus
+    cdef double complex lame = bulk_modulus - (2. / 3.) * shear_modulus
 
-    # Constants (See Eqs. B13-B16 of KMN15)
-    cdef double complex alpha2, beta2
-    cdef double gamma
-    alpha2 = (lame + 2. * shear_modulus) / density
-    beta2  = shear_modulus / density
-    gamma   = 4. * pi * G_to_use * density / 3.
+    # Constants
+    cdef double gamma          = 4. * pi * G_to_use * density / 3.
+    cdef double complex alpha2 = (lame + 2. * shear_modulus) / density
+    cdef double complex beta2  = shear_modulus / density
 
     # Optimizations
-    cdef double r_inverse, r2, degree_l_dbl, lp1, lm1, dlp1, llp1, dlp3
-    r_inverse    = 1. / radius
-    r2           = radius * radius
-    degree_l_dbl = <double> degree_l
-    lp1          = degree_l_dbl + 1.
-    lm1          = degree_l_dbl - 1.
-    dlp1         = 2. * degree_l_dbl + 1.
-    dlp3         = 2. * degree_l_dbl + 3.
-    llp1         = degree_l_dbl * lp1
+    cdef double r_inverse    = 1. / radius
+    cdef double r2           = radius * radius
+    cdef double degree_l_dbl = <double> degree_l
+    cdef double lp1          = degree_l_dbl + 1.
+    cdef double lm1          = degree_l_dbl - 1.
+    cdef double dlp1         = 2. * degree_l_dbl + 1.
+    cdef double dlp3         = 2. * degree_l_dbl + 3.
+    cdef double llp1         = degree_l_dbl * lp1
 
     # Helper functions
-    cdef double complex k2_quad_pos, k2_quad, k2_quad_sqrt
-    k2_quad_pos  = (4. * gamma / alpha2)
-    k2_quad      = k2_quad_pos**2 + ((4. * llp1 * gamma**2) / (alpha2 * beta2))
-    k2_quad_sqrt = cf_csqrt(k2_quad)
+    # See Eq. 99 of TS72
+    cdef double complex k2_quad_pos  = (4. * gamma / alpha2)
+    cdef double complex k2_quad      = k2_quad_pos**2 + ((4. * llp1 * gamma**2) / (alpha2 * beta2))
+    cdef double complex k2_quad_sqrt = cf_csqrt(k2_quad)
 
     # QUESTION: (Issue #43) TS74 has these flipped compared to KMN15. Going with TS74 for this func.
     # See the -/+ order in TS72 EQ. 99
     cdef size_t neg_index = 0
     cdef size_t pos_index = 1
-    cdef double complex k2_pos, k2_neg
-    k2_neg = (1. / 2.) * (k2_quad_pos - k2_quad_sqrt)
-    k2_pos = (1. / 2.) * (k2_quad_pos + k2_quad_sqrt)
+    cdef double complex k2_neg = (1. / 2.) * (k2_quad_pos - k2_quad_sqrt)
+    cdef double complex k2_pos = (1. / 2.) * (k2_quad_pos + k2_quad_sqrt)
 
-    cdef double complex f_pos, f_neg
-    f_pos = (beta2 * k2_pos) / gamma
-    f_neg = (beta2 * k2_neg) / gamma
+    cdef double complex f_pos = (beta2 * k2_pos) / gamma
+    cdef double complex f_neg = (beta2 * k2_neg) / gamma
 
-    cdef double complex h_pos, h_neg
-    h_pos = f_pos - lp1
-    h_neg = f_neg - lp1
+    cdef double complex h_pos = f_pos - lp1
+    cdef double complex h_neg = f_neg - lp1
 
     # Calculate Takeuchi and Saito functions
-    cdef double complex z2_pos, z2_neg
-    z2_pos = k2_pos * r2
-    z2_neg = k2_neg * r2
+    cdef double complex z2_pos = k2_pos * r2
+    cdef double complex z2_neg = k2_neg * r2
 
     cdef double complex phi_pos, phi_lp1_pos, psi_pos
     cdef double complex phi_neg, phi_lp1_neg, psi_neg
@@ -373,56 +356,52 @@ cdef void cf_takeuchi_liquid_dynamic_compressible(
 
     Parameters
     ----------
-    radius : float
-        Radius where the radial functions are calculated. [m]
-    bulk_modulus : Union[float, complex]
-        Bulk modulus (can be complex for dissipation) at `radius` [Pa]
-    density : float
-        Density at `radius` [kg m-3]
-    frequency : float
+    frequency : double
         Forcing frequency (for spin-synchronous tides this is the orbital motion) [rad s-1]
-    degree_l : int = 2
+    radius : double
+        Radius where the radial functions are calculated. [m]
+    density : double
+        Density at `radius` [kg m-3]
+    bulk_modulus : double complex
+        Bulk modulus (can be complex for dissipation) at `radius` [Pa]
+    degree_l : unsigned char
         Tidal harmonic order.
-    G_to_use : float = G
+    G_to_use : double
         Gravitational constant. Provide a non-dimensional version if the rest of the inputs are non-dimensional.
-
-    Returns
-    -------
-    solid_guesses : np.ndarray
-        The two independent liquid guesses (sn1, sn2)
-
+    num_ys : ssize_t
+        Number of radial solutions for this layer type.
+    starting_conditions_ptr : double complex*, <Output>
+        Desired starting conditions for this layer.
+        Two independent liquid guesses (sn1, sn2)
     """
 
     # Convert compressibility parameters
-    cdef double lame
-    lame = bulk_modulus
+    cdef double complex lame = bulk_modulus
 
-    # Constants (See Eqs. B13-B16 of KMN15)
-    cdef double dynamic_term, alpha2, gamma
-    dynamic_term = frequency * frequency
-    alpha2       = lame / density
-    gamma        = 4. * pi * G_to_use * density / 3.
+    # Constants
+    cdef double dynamic_term   = frequency * frequency
+    cdef double gamma          = 4. * pi * G_to_use * density / 3.
+    cdef double complex alpha2 = lame / density
 
     # Optimizations
-    cdef double r_inverse, r2, degree_l_dbl, lp1, lm1, dlp1, llp1, dlp3
-    r_inverse    = 1. / radius
-    r2           = radius * radius
-    degree_l_dbl = <double> degree_l
-    lp1          = degree_l_dbl + 1.
-    lm1          = degree_l_dbl - 1.
-    dlp1         = 2. * degree_l_dbl + 1.
-    dlp3         = 2. * degree_l_dbl + 3.
-    llp1         = degree_l_dbl * lp1
+    cdef double r_inverse    = 1. / radius
+    cdef double r2           = radius * radius
+    cdef double degree_l_dbl = <double> degree_l
+    cdef double lp1          = degree_l_dbl + 1.
+    cdef double lm1          = degree_l_dbl - 1.
+    cdef double dlp1         = 2. * degree_l_dbl + 1.
+    cdef double dlp3         = 2. * degree_l_dbl + 3.
+    cdef double llp1         = degree_l_dbl * lp1
 
     # Helper functions
-    cdef double k2, f, h
-    k2 = (1. / alpha2) * (dynamic_term + 4. * gamma - llp1 * gamma**2 / dynamic_term)
-    f  = -dynamic_term / gamma
-    h  = f - lp1
+    # k2 h and f no longer depend on k2. See Eq. 101 of TS72
+    cdef double f  = -dynamic_term / gamma
+    cdef double h  = f - lp1
+    cdef double complex k2 = (1. / alpha2) * (dynamic_term + 4. * gamma - llp1 * gamma**2 / dynamic_term)
 
     # Calculate Takeuchi and Saito functions
-    cdef double complex z, phi, phi_lp1, psi
-    z = k2 * r2
+    cdef double complex z = k2 * r2
+    cdef double complex phi, phi_lp1, psi
     cf_takeuchi_phi_psi(z, degree_l, &phi, &phi_lp1, &psi)
 
     # Found by setting mu=0 in Eq. 102 of TS72