+++ CHANGES.md
@@ -32,7 +32,7 @@ Fixes:
 
 Fixes:
 * Fixed dependency compatibility issues.
-* Fixed incorrect function signature type for scipy's `spherical_jn`. SciPy v.1.14.X uses a new signature which is breaking on MacOS. Limiting to "SciPy<1.14" for now.
+* Fixed incorrect function signature type for scipy's `spherical_jn`. SciPy v.1.14.X uses a new signature which is breaking on MacOS. Limiting to "SciPy<1.14" for now. See GitHub Issue #65
 
 ### Version 0.5.4 (2024-04-30)